requirements.txt
pytest>=7.0.0
pytest-allure-adaptor
pytest-rerunfailures
requests
allure-pytest
setuptools
python-dotenv